+++ src/lib.rs
@@ -3,5 +3,6 @@ pub mod collections;
 pub mod curl;
 pub mod models;
 pub mod net;
+pub mod postman;
 pub mod response;
 pub mod storage;